% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_week_aqiBar.R
\name{day_of_week_aqiBar}
\alias{day_of_week_aqiBar}
\title{Visualize the proportion of sensor measurements that fall within Air
Quality Index (AQI) categories for each hour of the given day.}
\usage{
day_of_week_aqiBar(sensor, aqi_country, day = "all", position = "stack")
}
\arguments{
\item{sensor}{An AirSensor object.}

\item{aqi_country}{The country who's Air Quality Index (AQI) should be used
for plotting, as an ISO 2 letter country code. Current options are "US" and
"IN".}

\item{day}{Character vector containing days of the week to show in plot (all
lowercase), or 'all' to use all days of the week.}

\item{position}{Whether to use stacked bar chart showing counts or filled bar
chart showing normalized proportions.}
}
\value{
a ggplot object.
}
\description{
Generate a bar chart of AQI measurements for a sensor.
}
