% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_ValidationPlot.R
\name{pat_ValidationPlot}
\alias{pat_ValidationPlot}
\title{Do sensor readings fall within spec?}
\usage{
pat_ValidationPlot(pat, param, param_lims = NULL)
}
\arguments{
\item{pat}{A Purple Air Time Series object.}

\item{param}{A Purple Air sensor parameter, one of "pm25", "temperature",
"humidity".}

\item{param_lims}{Numeric vector of length 2 specifying range of valid
measurements fpr plotting horizontal lines. If default parameters are based
off device specifications at
https://www2.purpleair.com/products/purpleair-pa-ii.}
}
\value{
ggplot object.
}
\description{
Generates a scatter plot of pat sensor readings for a specified
parameter with horizontal lines for valid reading limits.
}
\examples{
\dontrun{
  example_pat <- pat_load(label = "example")

  # Plotting with default specs.
  plot <- pat_ValidationPlot(pat = example_pat, param = 'pm25')

  # Plotting with custom specs.
  custom_lims <- c(5, 150)
  plot<- pat_ValidationPlot(pat = example_pat,
                            param = 'pm25',
                            param_lims = custom_lims)
}
}
