% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_hourlyavg_aqicolorsPlot.R
\name{sensor_hourlyavg_aqicolorsPlot}
\alias{sensor_hourlyavg_aqicolorsPlot}
\title{Sensor Hourly Averages Plot with points colored by Air Quality Index
Category}
\usage{
sensor_hourlyavg_aqicolorsPlot(
  sensor,
  aqi_country,
  time_facet = "none",
  point_size = 1,
  point_alpha = 0.5
)
}
\arguments{
\item{sensor}{AirSensor object.}

\item{aqi_country}{Character vector ISO 3166-1 alpha-2 country code of
country who's Air Quality Index (AQI) should be used for plotting
background. Options are United States ('US') or India ('IN').}

\item{time_facet}{The time unit by which to facet the plot, options are
'none', 'year', or 'month'. If 'month' faceting is done by both month and
year.}

\item{point_size}{The size of every point on the plot.}

\item{point_alpha}{An integer between 0 and 1, (inclusive), indicating the
opacity of each point on the plot.}
}
\value{
A ggplot object.
}
\description{
Input dataframe of PM hourly averages with columns datetime,
pm25, sensor, pm_channels_min, pm_channels_max, and a color palette with
length equal to the number of sensors in data. Generates a timeline plot of
hourly averages per sensor with maximum and minimum ribbons.
}
\references{
\code{citation("ggplot2")}
}
