% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workweek_weeknd_pmPlot.R
\name{workweek_weeknd_pmPlot}
\alias{workweek_weeknd_pmPlot}
\title{Plots Particulate Matter 2.5 Hour of Day Averages with lines showing
Workweek and Weekend Averages}
\usage{
workweek_weeknd_pmPlot(
  sensor_list,
  aqi_country,
  sensor_colors = NULL,
  aqi_bar_alpha = 0.2
)
}
\arguments{
\item{sensor_list}{A list of AirSensor objects.}

\item{aqi_country}{Character vector ISO 3166-1 alpha-2 country code of
country who's Air Quality Index (AQI) should be used for plotting
background. Current options are United States ("US") or India ("IN").}

\item{sensor_colors}{Character Vector the length of unique sensors in
outliercount_list, of hexadigit colors to use for distinguishing sensor
lines. If NULL, the default, a random palette will be produce. See details
for more information.}

\item{aqi_bar_alpha}{Numeric between 0 and 1 indicating the opacity of the
AQI background colors. Default is 0.2}
}
\value{
ggplot object.
}
\description{
Plot the mean PM2.5 value for each hour of the day, with
separate lines for weekdays and weekend values for one or more sensors.
}
